fix: remove app scrolling while viewing pdf and ppt

+++ lib/src/models/lesson_standing.dart
@@ -2,7 +2,7 @@ class LessonStanding {
   int number;
   List<int?> answers;
   int? score;
-  int progress;
+  double progress;
 
   LessonStanding(
       {required this.number,
